+++ samples/Omega.Sample.Car/Controllers/CarBookingService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Concurrent;
+using Servicecomb.Saga.Omega.Abstractions.Transaction;
 using Servicecomb.Saga.Omega.Core.Transaction;
 
 namespace Omega.Sample.Car.Controllers
