+++ samples/Omega.Sample.Hotel/Controllers/HotelBookingService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using Servicecomb.Saga.Omega.Abstractions.Transaction;
 using Servicecomb.Saga.Omega.Core.Transaction;
 
 namespace Omega.Sample.Hotel.Controllers
