+++ src/Servicecomb.Saga.Omega.Abstractions/Servicecomb.Saga.Omega.Abstractions.csproj
@@ -1,7 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="MethodBoundaryAspect.Fody" Version="1.0.66" />
+  </ItemGroup>
 
 </Project>
