+++ test/Servicecomb.Saga.Omega.Core.Tests/Context/OmegaContextTest.cs
@@ -15,6 +15,7 @@
  * limitations under the License.
  */
 
+using Servicecomb.Saga.Omega.Abstractions.Context;
 using Servicecomb.Saga.Omega.Core.Context;
 using Xunit;
